Refactor pipelines inside core library
