+++ paz/applications.py
@@ -7,3 +7,5 @@ from .pipelines import MiniXceptionFER
 from .pipelines import FaceKeypointNet2D32
 from .pipelines import HeadPoseKeypointNet2D32
 from .pipelines import HaarCascadeFrontalFace
+from .pipelines import RGBMaskToPowerDrillPose6D
+from .pipelines import PIX2POSEPowerDrill
